requirements.txt
numpy <= 1.19.5
keras <= 2.6.0
tensorflow <= 2.6.0
scikit-learn <= 0.24.2
librosa == 0.8.1
Pillow == 8.3.1
SoundFile
PySoundFile